Added Public Environment (#57)
This PR makes the `EnvironmentValues` API public, making the `environment(_:_:)` method available for all `Property` types. Currently, there are no reasons for developers to use this method mainly because the development of the `Environment` is still lacking, which allows access outside of internal methods as done today by RequestDL.

The purpose of this implementation is to pave the way for the implementation of other modifiers to enhance the implementation of other properties.

The implementation of `Environment` depends on the implementation of something similar to SwiftUI's `DynamicValue`, which is on the radar but has no release date yet.