Added environment configuration inside AnyProperty (#58)
When combining `AnyProperty` with another property that uses the environment, the environment values were recreated from scratch. This caused information to be lost and possible failures depending on the implementation.

With the changes made to the `Environment` API, it was possible to identify and fix this bug. With the implementation of the test, it was possible to reproduce this scenario and verify the fix.