Graph and Inputs refactor (#59)
We refactored `_GraphValue` to concentrate all operations and state verification of the property graph. To do this, we used the current implementation of `_PropertyInputs` as a basis.

After the refactoring, `_PropertyInputs` only contains input values that can be consumed during the construction of properties. An example of this is the current use of the environment to obtain environment values.

This change only has an internal effect and will enable the development of new features for RequestDL.