Implemented `@Namespace` (#65)
## Description

Implemented the `@Namespace` to integrate with planned changes that will
allow `Property` to store values using `@StoredObject` (not yet
available).

The purpose of `@Namespace` is to define a scope for all properties
within the object. This will help create references to objects and
prevent them from colliding, although this is unlikely.

Considering only this PR does not provide a complete view of the
changes.

Example:

```swift
struct MyCompanyConfiguration: Property {

    @Namespace var myCompany

    var body: some Property {
        Headers...
    }
}
```

With this, everything included by `MyCompanyConfiguration` will be
automatically identified internally as a property of the namespace
`_myCompany`. This will be leveraged by `@StoredObject`.

Fixes **None**

## Type of change

- [x] New feature (non-breaking change which adds functionality)

## Checklist

- [x] My code follows the code style of this project.
- [ ] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.