Added GraphID (#75)
## Description

We refactored _GraphValue to use GraphID and have greater control over
the hash that represents the current graph. This will help define the
rules for updating DynamicValue. As a result, the graph ID now
indirectly refers to the type of object it represents.

This structure may be subject to change in the future if new methods of
achieving the same goal are identified, or if new needs arise.

Fixes **None**

## Type of change

- [x] New feature (non-breaking change which adds functionality)

## Checklist

- [x] My code follows the code style of this project.
- [x] All new and existing tests passed.