Makes `HTTPClient` cached (#78)
## Description

This Pull Request (PR) introduces new functionality by implementing the
`Internals.Client` and `Internals.ClientManager` classes. These classes
are responsible for maintaining references to the `HTTPClient` and
`Internals.Session.Configuration` objects respectively.

The main objective of this implementation is to optimize the performance
of the `RequestDL` class by preventing it from repeatedly creating new
instances of objects related to `TLSConfiguration`. This operation can
be time-consuming as it involves reading multiple bytes.

Furthermore, this implementation takes advantage of the cached system
that is already included in the `Async-HTTP-Client`, which can help to
further improve the overall performance of the system.

Fixes **None**

## Type of change

- [x] New feature (non-breaking change which adds functionality)

## Checklist

- [x] My code follows the code style of this project.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.