Refactor URL configuration into baseURL, path, and queries (#84)
## Description

The current implementation of URL construction was limited to the
initial definition of the BaseURL and to Foundation's methods for URL
and URLComponents. This PR refactors the way the URL is constructed by
separating it into baseURL, paths, and queries. With this change, it is
possible to implement new features and make the existing ones more
stable.

Fixes **None**

## Type of change

- [x] New feature (non-breaking change which adds functionality)

## Checklist

- [x] My code follows the code style of this project.
- [x] All new and existing tests passed.