Move configuration of reading mode to request (#83)
## Description

With the Equatable update for the internal Session.Configuration, it was
discovered during the review of some methods that the ReadingMode was in
the wrong place. If more than one reading mode option was used in
different requests, it generated a bug that caused multiple instances of
the same HTTPClient. Therefore, this correction was labeled as a bug
fix.

Fixes **None**.

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] My code follows the code style of this project.
- [x] All new and existing tests passed.