Implemented URLEncoder (#85)
## Description

This PR implements URLEncoder to encode query values in the correct
format for transmission through URLs. The API was inspired by the
Foundation's JSONEncoder, despite some differences. The goal is to allow
developers to define their own methods for converting a value to a
String. With this implementation, developers will be able to easily
encode query values in their projects, ensuring that they are
transmitted correctly through URLs.

Fixes **None**

## Type of change

- [x] New feature (non-breaking change which adds functionality)

## Checklist

- [x] My code follows the code style of this project.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [x] I have added/updated necessary comments/documentation.