Internals.ByteURL improvements (#88)
## Description

A new feature has been added to ByteURL, allowing it to be instantiated
with a reference to a ByteBuffer. This change enables the direct use of
DataBuffer by DownloadBuffer, replacing the previous usage of
ByteBuffer. As a result, byte manipulation has become more robust and
consistent with the rest of the project, building upon the existing
optimizations.

Furthermore, a bug related to the use of buffers with a length greater
than their storage capacity has been addressed. Instead of triggering a
fatal error, the buffers now gracefully return nil.

Fixes **None**

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] My code follows the code style of this project.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.