Resolved implementation (#103)
## Description

This PR includes internal changes to improve the process of resolving
properties in the `Internal.Session` and `Internal.Request` components.
The updates aim to enhance the overall functionality and efficiency of
the codebase.

Previously, the result of the property resolution process in
`Internal.Session` and `Internal.Request` was returned directly.
However, in this PR, we have made internal modifications so that the
`Resolve` method now returns a `Resolved` object. This `Resolved` object
contains all the necessary information required to execute the request
effectively.

By introducing the `Resolved` object, we centralize and encapsulate the
relevant data, ensuring a more structured and comprehensive approach to
property resolution. This change simplifies the code logic and improves
code readability by providing a single object that holds all the
pertinent information.

These modifications have been implemented to enhance the overall
efficiency, maintainability, and extensibility of the codebase. They
provide a clear and concise way to access and utilize the resolved
properties, streamlining the execution of requests within the
`Internal.Session` and `Internal.Request` components.

Please note that these changes do not impact the external behavior or
functionality of the code. They solely focus on improving the internal
process of resolving properties, resulting in a more robust and
streamlined codebase.

Fixes **None**

## Type of change

- [x] New feature (non-breaking change which adds functionality)

## Checklist

- [x] My code follows the code style of this project.
- [x] All new and existing tests passed.