AsyncStream implementation (#108)
## Description

After the implementation of Sendable, a problem was identified in the
current implementation of DataStream. It was built using a combination
of queuing and closure concepts.

However, this solution is useful only when we have a single object as
the consumer of DataStream. With the implementation of caching, the data
flow is fragmented as the cache independently consumes the data to save
locally. Therefore, given this scenario, it was necessary to reconstruct
AsyncStream using a more robust set of objects, leveraging the concepts
of Sendable and locks.

Fixes **None**

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.