Cache System (#107)
## Description

We are excited to announce that we have completed the development of the
Cache feature with several interesting enhancements. We have introduced
three new methods to configure cache usage in your requests:
`cacheStrategy(_:)`, `cache(memoryCapacity:diskCapacity:)`, and
`cachePolicy(_:)`.

The `CacheStrategy` indicates how RequestDL should query the stored
cache and return the data to the application. We offer a range of
options, from not using the cache at all to exclusively relying on
cached data.

The `CachePolicy` informs RequestDL whether the request should have its
cache stored only in memory, on disk, or both.

Regarding the `Cache` (DataCache) itself, it allows you to specify the
cache system to be used. The option with just
`.init(memoryCapacity:diskCapacity:)` will utilize the identifier of the
`Bundle.main` to save the cache. The second option involves the
`suitName:` parameter, which serves as the identifier used for caching.
The last option involves the `url:` parameter, enabling you to define
the exact location where the cache should be saved.

My todo list:

- [x] Verify cache expiration using headers
- [x] ~Lock cache while is writing block attempt to write over current
process~
- [x] Return cache only if it was totally wrote
- [x] Cache responses with zero bytes

Fixes **None**

## Type of change

- [x] New feature (non-breaking change which adds functionality)

## Checklist

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.