Add RequestTaskModifier (#128)
## Description

<!--- 
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context.
-->

This MR introduces the brand new `RequestTaskModifier` that offers an
improved syntax for creating custom modifiers. Previously, it had a
dependency on the Body task that needed modification, but now it has
been decoupled from it. The changes now specifically define the Input
and Output for the element only.

You should implement this if you want to modify the Input to Output.

```swift
struct ExampleModifier<Input, Output>: RequestTaskModifier {

    func body(_ task: Content) -> Output {}
}
```

Alternatively, if you prefer to maintain the original output:

```swift
struct ExampleModifier<Input>: RequestTaskModifier {

    func body(_ task: Content) -> Input {}
}
```

Fixes **None**

## Type of change

<!--- Please delete options that are not relevant. -->

- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist

<!--- Please delete options that are not relevant. -->

- [x] My code follows the code style of this project.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.