BaseURL Exceptions Removal (#141)
## Description

<!--- 
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context.
-->

An issue that arose during the implementation of `2.*.*` is the
excessive use of fatal errors, which makes the use of RequestDL unstable
with the possibility of critical failures in production. This PR is part
of a series of PRs aimed at removing fatal errors and using `throw` when
such a possibility exists.

Fixes **None**

## Type of change

<!--- Please delete options that are not relevant. -->

- [x] Bug fix (non-breaking change which fixes an issue)

## Checklist

<!--- Please delete options that are not relevant. -->

- [x] My code follows the code style of this project.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.