Improved `Progress` protocol (#143)
## Description

<!--- 
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context.
-->

This pull request addresses a series of issues related to the current
implementation of the `Progress` protocol.

The first problem addressed was the lack of a parameter in
`UploadProgress` to determine the total number of bytes being sent to
the server. This limitation rendered the progress implementation
significantly limited and purposeless.

The second problem involved a major refactoring of `DownloadProgress`,
which previously required specifying the header key to determine the
total number of bytes the client would receive from the server. The
logic has now been internalized and utilizes the `Content-Length`
header, which is the standard used by AsyncHTTPClient.

The third problem involved how RequestDL internally tracked the total
number of bytes sent and received. The fix introduced several new
properties that provide accurate information about the size of the data
being sent and received.

This PR has some breaking changes:

- UploadProgress
  ```swift
  // Before
  protocol UploadProgress {
  
      func upload(_ bytesLength: Int)
  }
  
  // After
  protocol UploadProgress {
  
      func upload(_ chunkSize: Int, totalSize: Int)
  }
  ```

- DownloadProgress
  ```swift
  // Before
  protocol DownloadProgress {
  
      var contentLengthHeaderKey: String? { get }
      
      func download(_ part: Data, length: Int?)
  }
  
  // After
  protocol DownloadProgress {
  
      func download(_ slice: Data, totalSize: Int)
  }
  ```

- Response
  ```swift
  // Before
  enum Response {
      case upload(Int)
      case download(ResponseHead, AsyncBytes)
  }
  
  // After
  enum ResponseStep {
      case upload(UploadStep)
      case download(DownloadStep)
  }
  ```

Fixes **None**

## Type of change

<!--- Please delete options that are not relevant. -->

- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist

<!--- Please delete options that are not relevant. -->

- [x] My code follows the code style of this project.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [x] I have added/updated necessary comments/documentation.