New workflow for code coverage and tests (#224)
## Description

<!--- 
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context.
-->

This PR enhances the code coverage collection process. Previously, the
code was only compiled for macOS, but the reality is that the RequestDL
is available for multiple platforms. Therefore, the workflows now align
with this and test the conformance of all platforms.

---------

Co-authored-by: brennobemoura <37243584+brennobemoura@users.noreply.github.com>