Refactor Locks and TaskEnvironmentValues for better usage (#226)
## Description

<!--- 
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context.
-->

This enhancement improves the usage of Lock. The primary objective is to
enhance the usage of TaskEnvironment to make it public and accessible
using the `@TaskEnvironment` decorator. The subsequent updates will make
the `TaskEnvironmentValues` and `PropertyEnvironmentValues` the same
object.

## Type of change

<!--- Please delete options that are not relevant. -->

- [x] New feature (non-breaking change which adds functionality)

## Checklist

<!--- Please delete options that are not relevant. -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [x] I have added/updated necessary comments/documentation.

Co-authored-by: brennobemoura <37243584+brennobemoura@users.noreply.github.com>