Implements Logger Supports (#233)
## Description

<!--- 
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context.
-->

This PR introduces support for Logger. Initially, the proposal aimed to
include OSLog support, but due to the Multiplatform nature of the
project, we can only provide Swift-Log support for now. In the future,
we may be able to expand this to support OSLog as well.

Fixes #195

## Type of change

<!--- Please delete options that are not relevant. -->

- [x] New feature (non-breaking change which adds functionality)

## Checklist

<!--- Please delete options that are not relevant. -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [x] I have added/updated necessary comments/documentation.

---------

Co-authored-by: brennobemoura <37243584+brennobemoura@users.noreply.github.com>