Fix invalid state in AsyncLock (#234)
## Description

<!--- 
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context.
-->

The AsyncLock was entering an invalid state when a locked task was
canceled. This occurred because the unlock operation failed to skip the
non-pending tasks, resulting in an error of invalid state that persisted
indefinitely.

## Type of change

<!--- Please delete options that are not relevant. -->

- [x] Bug fix (non-breaking change which fixes an issue)

Co-authored-by: brennobemoura <37243584+brennobemoura@users.noreply.github.com>